#113 Moved SaaS tests to a separate test set
