+++ .github/workflows/ci.yml
@@ -17,10 +17,6 @@ jobs:
 
   ci-job:
     name: Checks
-    env:
-      SAAS_HOST: ${{ secrets.INTEGRATION_TEAM_SAAS_STAGING_HOST }}
-      SAAS_ACCOUNT_ID: ${{ secrets.INTEGRATION_TEAM_SAAS_STAGING_ACCOUNT_ID }}
-      SAAS_PAT: ${{ secrets.INTEGRATION_TEAM_SAAS_STAGING_PAT }}
     uses: exasol/python-toolbox/.github/workflows/checks.yml@0.3.0
 
   tests-job:
@@ -53,3 +49,10 @@ jobs:
 
       - name: Run Tests
         run: poetry run pytest tests
+
+      - name: Run SaaS Tests
+        env:
+          SAAS_HOST: ${{ secrets.INTEGRATION_TEAM_SAAS_STAGING_HOST }}
+          SAAS_ACCOUNT_ID: ${{ secrets.INTEGRATION_TEAM_SAAS_STAGING_ACCOUNT_ID }}
+          SAAS_PAT: ${{ secrets.INTEGRATION_TEAM_SAAS_STAGING_PAT }}
+        run: poetry run pytest test_saas
