+++ .github/workflows/pr_merge.yml
@@ -10,10 +10,6 @@ jobs:
 
   ci-job:
     name: Checks
-    env:
-      SAAS_HOST: ${{ secrets.INTEGRATION_TEAM_SAAS_STAGING_HOST }}
-      SAAS_ACCOUNT_ID: ${{ secrets.INTEGRATION_TEAM_SAAS_STAGING_ACCOUNT_ID }}
-      SAAS_PAT: ${{ secrets.INTEGRATION_TEAM_SAAS_STAGING_PAT }}
     uses: exasol/python-toolbox/.github/workflows/checks.yml@0.3.0
 
   tests-job:
