+++ test/integration/test_bucket_path.py
@@ -109,59 +109,3 @@ def test_write_delete_onprem(test_config, children_poem, classic_poem):
                 test_config.password,
                 str(poem_path)
             )
-
-
-def test_write_read_back_saas(saas_test_service_url, saas_test_token,
-                              saas_test_account_id, saas_test_database_id,
-                              children_poem):
-
-    base_path = bfs.path.build_path(backend=bfs.path.StorageBackend.saas,
-                                    url=saas_test_service_url,
-                                    account_id=saas_test_account_id,
-                                    database_id=saas_test_database_id,
-                                    pat=saas_test_token)
-    file_name = 'test_bucket_path/test_write_read_back_saas/little_star.txt'
-    poem_path = base_path / file_name
-
-    poem_path.write(children_poem)
-    data_back = b''.join(poem_path.read(20))
-    assert data_back == children_poem
-
-
-def test_write_list_files_saas(saas_test_service_url, saas_test_token,
-                               saas_test_account_id, saas_test_database_id,
-                               children_poem, classic_poem):
-
-    base_path = bfs.path.build_path(backend=bfs.path.StorageBackend.saas,
-                                    url=saas_test_service_url,
-                                    account_id=saas_test_account_id,
-                                    database_id=saas_test_database_id,
-                                    pat=saas_test_token,
-                                    path='test_bucket_path/test_write_list_files_saas')
-    poem_path1 = base_path / 'children/little_star.txt'
-    poem_path2 = base_path / 'classic/highlands.txt'
-
-    poem_path1.write(children_poem)
-    poem_path2.write(classic_poem)
-    expected_names = {'children', 'classic', 'little_star.txt', 'highlands.txt'}
-    assert _collect_all_names(base_path) == expected_names
-
-
-def test_write_delete_saas(saas_test_service_url, saas_test_token,
-                           saas_test_account_id, saas_test_database_id,
-                           children_poem, classic_poem):
-
-    base_path = bfs.path.build_path(backend=bfs.path.StorageBackend.saas,
-                                    url=saas_test_service_url,
-                                    account_id=saas_test_account_id,
-                                    database_id=saas_test_database_id,
-                                    pat=saas_test_token)
-    poems_root = base_path / 'test_bucket_path/test_write_delete_saas'
-    poem_path1 = poems_root / 'children/little_star.txt'
-    poem_path2 = poems_root / 'classic/highlands.txt'
-
-    poem_path1.write(children_poem)
-    poem_path2.write(classic_poem)
-    poem_path1.rm()
-    expected_names = {'classic', 'highlands.txt'}
-    assert _collect_all_names(poems_root) == expected_names
